Create ChaserGame3
The original start of this game had significant issues with trying to create a new chaser. I couldn't find the solution to find i to be resolved to a variable. I then figured out how to update and just using the if loop if( !unvulnerable&& c. is Active ()&& c. hits(mouseX, mouseY). Then it started to work. 
I also used the invulnerability in the draw.