Add files via upload
I tried adding a super form of Messi in that code using a PImage, but I couldn't add splats to the chaser or the players. I pulled each splat of the chasers where they were needed. I placed different classes, and setups, and drew in their own separate courses I couldn't find a way to remove the splat.