Changes
- Changed the game state variable to an initial 0 (title screen). The code should move into state one (intro), where the start sound plays for 5 seconds. Once 5 seconds have passed, the game enters state 2 (game running), and the player is instantiated.
- Super Form Swap:
I tried to add to the player class to the point  that a super image and normal image are both accepted. when the global invulnerability flag is true ( which you already use the meter trigger), the player's display method will draw the super image and goes back to normal when invulnerability ends.
-Splats are now used by using the size depending character.
I couldn't figure out the syntax error before finishing  near mouse pressed.